[mod/#50] Record 중첩 네비 제거
